working on below reference non incentized advancements, checking tx status in model

+++ model/chain/run.sh
@@ -11,8 +11,10 @@ set -e
 # Clean it out
 echo "Cleaning Old Database..."
 mkdir -p ./empty_db
+#: '
 time rsync -a --delete ./empty_db/ ./db/
 rm -Rf db go-ethereum-keystore*
+#'
 
 # Have a function to kill off Ganache and clean up the database when we quit.
 function cleanup {
@@ -23,20 +25,28 @@ function cleanup {
     # Clean database
     echo "Cleaning Up Database..."
     mkdir -p ./empty_db
-    time rsync -a --delete ./empty_db/ ./db/
-    rm -Rf db go-ethereum-keystore*
+    if [[ "${SAVE_STATE}" == "1" ]] ; then
+        echo "Just Kidding... Leaving Database Alone"
+    else
+        time rsync -a --delete ./empty_db/ ./db/
+        rm -Rf db go-ethereum-keystore*
+        rm *-approvals.json
+    fi
 }
 
 trap cleanup EXIT
 
+: '
+# THIS IS NEEDED TO CHECK FOR DIFFS BETWEEN UPDATES
+diff $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.26/plugin/evm/client.go $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.27/plugin/evm/client.go
+diff $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.26/plugin/evm/service.go $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.27/plugin/evm/service.go
+diff $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.26/miner/worker.go $GOPATH/pkg/mod/github.com/ava-labs/coreth@\@v0.3.27/miner/worker.go
+'
+
 # Start the chain
-#echo "Starting Ganache..."
 # Need to run the below command in a while loop when deploying locally
-# curl -H "Content-Type: application/json" -X POST --data '{"id":1337,"jsonrpc":"2.0","method":"evm_mine","params":[]}' http://localhost:7545
 echo "Starting AvalancheGo..."
 TMPDIR="$(pwd)" avalanchego --network-id=local --staking-enabled=false --snow-sample-size=1 --snow-quorum-size=1 --db-dir=./db/ --http-port=9545 --log-level=verbo --log-dir=./db/ --snow-avalanche-batch-size=1 > ganache_output.txt &
-#TMPDIR="$(pwd)" ganache-cli --p 7545 --gasLimit 8000000 --accounts 20 --defaultBalanceEther 1000000 --db ./db --noVMErrorsOnRPCResponse > ganache_output.txt &
-#TMPDIR="$(pwd)" ganache-cli --p 7545 --gasLimit 8000000 --accounts 20 --defaultBalanceEther 1000000 --blockTime 604800 --db ./db --noVMErrorsOnRPCResponse  >ganache_output.txt &
 GANACHE=$!
 
 # Wait for it to come up
@@ -46,7 +56,6 @@ echo "Waiting for AvalancheGo..."
 while ! grep -i "listening on" ganache_output.txt 2>/dev/null ; do
     sleep 1
 done
-
 #: '
 # Creating accounts
 echo "Creating deploy test accounts..."
@@ -55,7 +64,7 @@ time truffle exec make_accounts.js --network development > make_accounts_output.
 echo "Deploying contracts..."
 time truffle migrate --reset --skip-dry-run --network=development | tee deploy_output.txt
 echo "Creating sim test accounts..."
-time truffle exec make_accounts.js --network development --max-accounts 40 >> make_accounts_output.txt
+time truffle exec make_accounts.js --network development --max-accounts 20 >> make_accounts_output.txt
 
 #truffle migrate --reset --network=development | tee deploy_output.txt
 #'
